usando stdint

+++ teste.ino
@@ -1,10 +1,10 @@
-int AC_1 = 32;
-unsigned long counter_LOW = 0;
-unsigned long counter_HIGH = 0;
+uint8_t AC_1 = 32;
+uint32_t counter_LOW = 0;
+uint32_t counter_HIGH = 0;
 
-unsigned long previousMillis = 0;
+uint32_t previousMillis = 0;
 
-const long interval = 1000; // Esse será o intervalo de contagem
+const uint16_t interval = 1000; // Esse será o intervalo de contagem
 
 void setup() {
   Serial.begin(9600);
@@ -12,7 +12,7 @@ void setup() {
 }
 
 void loop() {
-  unsigned long currentMillis = millis();  
+  uint32_t currentMillis = millis();  
 
   counter_LOW = 0;
   counter_HIGH = 0;
@@ -20,7 +20,7 @@ void loop() {
   while (currentMillis - previousMillis <= interval) {
     previousMillis = currentMillis;
     
-    byte VAC_1 = digitalRead(AC_1);    
+    uint8_t VAC_1 = digitalRead(AC_1);    
        
     if (!VAC_1){
       counter_LOW++;
